refactor(store): split terminalRegistrySlice by domain
Split monolithic terminalRegistrySlice.ts (~1708 lines) into focused domain modules:

- Extract types to types.ts (interfaces, type aliases)
- Extract helpers to helpers.ts (constants, deriveRuntimeStatus, getDefaultTitle)
- Extract persistence logic to persistence.ts (saveTerminals, flushTerminalPersistence)
- Extract trash/restore logic to trash.ts (createTrashExpiryHelpers)
- Extract layout optimization to layout.ts (optimizeForDock)
- Compose all actions in index.ts with re-exports for backward compatibility
- Maintain backward compatibility via re-export layer in terminalRegistrySlice.ts
- Add TerminalRegistryStoreApi to public exports

All store behavior unchanged. Tests pass. No breaking changes.
